+++ ios/LibVlcPlayerView.swift
@@ -390,7 +390,7 @@ class LibVlcPlayerView: ExpoView {
                     player.time = VLCTime(int: Int32(value))
                 }
             } else {
-                time = Int(Int32(value) * mediaLength)
+                time = Int(value * Double(mediaLength))
             }
         }
     }
